Major core optimizations, fixes, and housekeeping. Fixed compile output path for uoAvos project settings. Fixed a bug that would skip home town search for some vendors. Added new FontData asset manager. Added new region-based harvest nodes. Added new Quest system, pending example quests. Added new currency management options to vendors. Added new utility helper methods for common tasks. Added new gump functions for generating tabs and scrollbar controls. Cleaned boats in preparation for new collision detection API. Revert to traditional save file path to avoid issues with external user code.

+++ Scripts/Content/System/Guild/BaseGuild.cs
@@ -2712,7 +2712,7 @@ namespace Server.Guilds
 			AddBackground(450, 370, 100, 26, 0x2486);
 
 			AddButton(455, 375, 0x845, 0x846, 7, GumpButtonType.Reply, 0);
-			AddHtmlLocalized(480, 373, 60, 26, 3006115, (m_IsResigning) ? 0x5000 : 0, false, false); // Resign
+			AddHtmlLocalized(480, 373, 60, 26, 3006115, (short)(m_IsResigning ? 0x5000 : 0), false, false); // Resign
 		}
 
 		public override void OnResponse(NetState sender, RelayInfo info)
@@ -2867,7 +2867,7 @@ namespace Server.Guilds
 
 			AddBackground(20, 180, 150, 26, 0x2486);
 			AddButton(25, 185, 0x845, 0x846, 1, GumpButtonType.Reply, 0);
-			AddHtmlLocalized(50, 183, 110, 26, 1062993, (m_ToLeader) ? 0x990000 : 0, false, false); // Promote
+			AddHtmlLocalized(50, 183, 110, 26, 1062993, (short)(m_ToLeader ? 0x990000 : 0), false, false); // Promote
 
 			AddBackground(180, 180, 150, 26, 0x2486);
 			AddButton(185, 185, 0x845, 0x846, 3, GumpButtonType.Reply, 0);
@@ -2879,7 +2879,7 @@ namespace Server.Guilds
 
 			AddBackground(180, 210, 150, 26, 0x2486);
 			AddButton(185, 215, 0x845, 0x846, 5, GumpButtonType.Reply, 0);
-			AddHtmlLocalized(210, 213, 110, 26, 1062997, (m_toKick) ? 0x5000 : 0, false, false); // Kick
+			AddHtmlLocalized(210, 213, 110, 26, 1062997, (short)(m_toKick ? 0x5000 : 0), false, false); // Kick
 		}
 
 		public override void OnResponse(NetState sender, RelayInfo info)
